+++ tests/framework/db/sqlite/SqliteConnectionTest.php
@@ -47,6 +47,8 @@ class SqliteConnectionTest extends ConnectionTest
         $this->assertEquals("`table`.`column`", $connection->quoteColumnName('table.column'));
         $this->assertEquals('[[column]]', $connection->quoteColumnName('[[column]]'));
         $this->assertEquals('{{column}}', $connection->quoteColumnName('{{column}}'));
+        $this->assertEquals('{{table}}.`column`', $connection->quoteColumnName('{{table}}.column'));
+        $this->assertEquals('{{table}}.[[column]]', $connection->quoteColumnName('{{table}}.[[column]]'));
         $this->assertEquals('(column)', $connection->quoteColumnName('(column)'));
     }